+++ tests/test_casing.py
@@ -1,8 +1,15 @@
 from dataclasses import asdict, dataclass
+from typing import Optional
 
 from quart import Quart
 
-from quart_schema import QuartSchema, ResponseReturnValue, validate_request, validate_response
+from quart_schema import (
+    QuartSchema,
+    ResponseReturnValue,
+    validate_querystring,
+    validate_request,
+    validate_response,
+)
 
 
 @dataclass
@@ -36,3 +43,24 @@ async def test_response_casing() -> None:
     test_client = app.test_client()
     response = await test_client.get("/")
     assert await response.get_data(as_text=True) == '{"snakeCase":"Hello"}'
+
+
+@dataclass
+class QueryData:
+    snake_case: Optional[str] = None
+
+
+async def test_querystring_casing() -> None:
+    app = Quart(__name__)
+    QuartSchema(app, convert_casing=True)
+
+    @app.get("/")
+    @validate_querystring(QueryData)
+    async def index(query_args: QueryData) -> ResponseReturnValue:
+        return str(asdict(query_args))
+
+    test_client = app.test_client()
+    response = await test_client.get("/", query_string={"snake_case": "Hello"})
+    assert await response.get_data(as_text=True) == "{'snake_case': 'Hello'}"
+    response = await test_client.get("/?snakeCase=Hello")
+    assert await response.get_data(as_text=True) == "{'snake_case': 'Hello'}"