+++ tests/test_openapi.py
@@ -1,10 +1,12 @@
 from dataclasses import dataclass
-from typing import Optional, Tuple
+from typing import Dict, List, Optional, Tuple
 
+from pydantic import Field
 from quart import Quart
 
 from quart_schema import (
     QuartSchema,
+    security_scheme,
     validate_headers,
     validate_querystring,
     validate_request,
@@ -14,7 +16,7 @@ from quart_schema import (
 
 @dataclass
 class QueryItem:
-    count_le: Optional[int] = None
+    count_le: Optional[int] = Field(description="count_le description")
 
 
 @dataclass
@@ -30,7 +32,7 @@ class Result:
 
 @dataclass
 class Headers:
-    x_name: str
+    x_name: str = Field(..., description="x-name description")
 
 
 async def test_openapi() -> None:
@@ -43,7 +45,20 @@ async def test_openapi() -> None:
     @validate_headers(Headers)
     @validate_response(Result, 200, Headers)
     async def index() -> Tuple[Result, int, Headers]:
+        """Summary
+        Multi-line
+        description.
+
+        This is a new paragraph
+
+            And this is an indented codeblock.
+
+        And another paragraph."""
         return Result(name="bob"), 200, Headers(x_name="jeff")
+
+    @app.websocket("/ws")
+    async def ws() -> None:
+        pass
 
     test_client = app.test_client()
     response = await test_client.get("/openapi.json")
@@ -54,15 +69,20 @@ async def test_openapi() -> None:
         "paths": {
             "/": {
                 "get": {
+                    "summary": "Summary",
+                    "description": "Multi-line\ndescription.\n\nThis is a new paragraph\n\n    "
+                    "And this is an indented codeblock.\n\nAnd another paragraph.",
                     "parameters": [
                         {
                             "in": "query",
                             "name": "count_le",
+                            "description": "count_le description",
                             "schema": {"title": "Count Le", "type": "integer"},
                         },
                         {
                             "in": "header",
                             "name": "x-name",
+                            "description": "x-name description",
                             "schema": {"title": "X Name", "type": "string"},
                         },
                     ],
@@ -93,15 +113,73 @@ async def test_openapi() -> None:
                                     }
                                 },
                                 "headers": {
-                                    "x-name": {"schema": {"title": "X Name", "type": "string"}}
+                                    "x-name": {
+                                        "schema": {
+                                            "title": "X Name",
+                                            "type": "string",
+                                            "description": "x-name description",
+                                        }
+                                    }
                                 },
                             },
-                            "description": "Result(name: str)",
+                            "description": "",
                         }
                     },
                 }
             }
         },
-        "servers": [],
-        "tags": [],
     }
+
+
+async def test_security_schemes() -> None:
+    app = Quart(__name__)
+    QuartSchema(
+        app,
+        security_schemes={
+            "MyBearer": {"type": "http", "scheme": "bearer"},
+            "MyBasicAuth": {"type": "http", "scheme": "basic"},
+        },
+        security=[{"MyBearer": []}, {"MyBasicAuth": ["foo", "bar"]}],
+    )
+
+    @app.route("/")
+    @security_scheme([{"MyBearer": []}])
+    async def index() -> Tuple[Dict, int]:
+        return {}, 200
+
+    test_client = app.test_client()
+    response = await (await test_client.get("/openapi.json")).get_json()
+    assert response["security"] == [{"MyBearer": []}, {"MyBasicAuth": ["foo", "bar"]}]
+    assert response["components"]["securitySchemes"] == {
+        "MyBearer": {"type": "http", "scheme": "bearer"},
+        "MyBasicAuth": {"type": "http", "scheme": "basic"},
+    }
+    assert response["paths"]["/"]["get"]["security"] == [{"MyBearer": []}]
+
+
+@dataclass
+class Employee:
+    name: str
+
+
+@dataclass
+class Employees:
+    resources: List[Employee]
+
+
+async def test_openapi_refs() -> None:
+    app = Quart(__name__)
+    QuartSchema(app, convert_casing=True)
+
+    @app.route("/")
+    @validate_response(Employees)
+    async def index() -> Employees:
+        return Employees(resources=[Employee(name="bob")])
+
+    test_client = app.test_client()
+    response = await test_client.get("/openapi.json")
+    schema = await response.get_json()
+    ref = schema["paths"]["/"]["get"]["responses"]["200"]["content"]["application/json"]["schema"][
+        "properties"
+    ]["resources"]["items"]["$ref"]
+    assert ref[len("#/components/schemas/") :] in schema["components"]["schemas"].keys()