+++ CHANGELOG.rst
@@ -1,3 +1,21 @@
+0.13.0 2022-09-04
+-----------------
+
+* Make it clear that redocs require Javascript to work.
+* Add the ability to add security tags.
+* Ensure the tag decorator overwrites any existing, this is a
+  backwards incompatible change.
+* Ensure casing conversion happens for request query string args.
+
+0.12.0 2022-07-23
+-----------------
+
+* Increase swagger version to 4.12.0.
+* Render endpoint summary not as code.
+* Bugfix ignore websocket routes when generating openapi description.
+* Require Quart >= 0.18
+* Switch to GitHub rather than GitLab.
+
 0.11.1 2022-02-03
 -----------------